requirements.txt
fastapi>=0.115.2
uvicorn>=0.22.0
gradio>=4.0.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
scipy>=1.10.0